added pages folder
This folder contains the sheets of the dashboard, powered by the app.py and utils.py files. They have the layout and the callbacks of each graph and functionality for that page. 

NOTE: As of 12 December 2025, the pass by url is being updated to incorporate the contractor_id link for the user (In progress). Also the UI is a bit clogged for the attendance and payroll sheets. Currently working on it.